requirements.txt
flask
ansible
ansible_runner
pycryptodome